--- src/index.ts
@@ -1,6 +1,0 @@
-export * from './core/index.js';
-export * from './core/balancer.js';
-export * from './core/inspector.js';
-
-export { default as curry } from './core/lambda/curry.js';
-export { default as pipe } from './core/lambda/pipe.js';
